refactor: constants 상수들 파일 분리 및 라우팅 경로 상수화
