+++ src/constant/common.ts
@@ -1,0 +1,4 @@
+export const COUPON_TYPE_UNIT = {
+  percent: '%',
+  amount: '원',
+};
