+++ src/constant/user.ts
@@ -1,0 +1,4 @@
+const TEMP_ID = 'a@a.com';
+const TEMP_PW = '1234';
+
+export const USER_AUTH_TOKEN = btoa(`${TEMP_ID}:${TEMP_PW}`);
