+++ src/globalState/atoms/serverName.ts
@@ -1,6 +1,6 @@
 import { atom } from 'recoil';
 import { ServerName } from '../../types/server';
-import { DEFAULT_SERVER_NAME } from '../../constant';
+import { DEFAULT_SERVER_NAME } from '../../constant/server';
 
 const serverNameState = atom<ServerName>({
   key: 'serverNameState',
