+++ src/mocks/handlers.ts
@@ -9,7 +9,7 @@ import {
   ORDERS_PATH_NAME,
   PRODUCTS_PATH_NAME,
   USERS_COUPON_PATH_NAME,
-} from '../constant';
+} from '../constant/server';
 import LocalStorage from '../utils/LocalStorage';
 import type { CartProduct, Coupon, Order } from '../types/product';
